Implement refresh timeout outside provider
The Cloudscale provider implementation would use a context with
a timeout on its own. With the upcoming Exoscale provider implementation
it became clear that the refresh timeout should be applied outside the
provider implementation (with a context with a timeout). With this
change the global configuration object no longer needs to be given to
the provider implementation too.