Pass logger to refresher
With the upcoming Exoscale provider implementation there will also be
a need for an address-specific logger in its refresher. Considering that
it's the same as for Cloudscale we can construct it outside the provider
and pass it in. At the same time the configuration doesn't have to be
passed as it's not used.