Add Debian and RPM packaging
So far the static Ursula binary was only packaged in a Docker container. With
this change FPM[1] is used to also create Debian and RPM packages. These
packages don't conform to all standards, i.e. they contain no changelog. The
alternative of building full-fledged packages was considered too elaborate as
it would require providing Go and dependencies on the respective systems.

The build chain also includes an image for FPM. That will be moved to
a separate location with time.

[1] http://fpm.readthedocs.io