Longer timeout, log errors during Exoscale metadata discovery
When the discovery of the Exoscale instance metadata failed only the
last failure was returned. For debugging it may be helpful to be able to
get more detailed logs. Write all errors as debug messages.

The timeout for retrieving metadata is doubled from 5 to 10 seconds.
While a longer outage still leads to a permanent failure it gives the
process more time to retrieve the metadata.