Use Cloudscale meta service to discover server UUID
The Cloudscale API identifies servers by their unique ID, a UUID. So far
it was necessary to either specify an explicit UUID ("server-uuid") or
a map from hostname to UUID ("hostname-to-server-uuid"). In July 2018
Cloudscale have started to export the server UUID via the OpenStack
metadata service.

Just like Exoscale the code tries to discover the UUID upon start and
implements retries for failures.