Support specifying managed addresses in config
Originally Ursula would always use IP addresses found in the Keepalived
configuration. In certain configuration scenarios there are none, i.e.
if only a program such as Ursula is meant to be invoked on state
changes. With this change it becomes possible to list addresses to be
managed in the configuration file. They override any found in the
Keepalived configuration.