Terminate after Keepalived process has gone away
Keepalived does not terminate long-running notification programs when
exiting. In addition Keepalived may be terminated through other means
such as SIGKILL. In such cases the IP address updates must stop as soon
as possible. As of Keepalived 1.2, shipped with OpenShift 3.9, there is
no mechanism to reliably detect that Keepalived has terminated. Later
versions have support for FIFOs to communicate to notification programs.
Therefore the only reasonable method is to locate the process ID of
Keepalived and polling for its validity in a regular interval.