Use function to compare UUIDs
The update of the exoscale/egoscale library in commit b366784 changed
the type of instance and NIC identifiers from strings to
"*egoscale.UUID". In one case the comparison of UUIDs was effectively
a comparison of pointers. The result was flapping of floating IP
addresses because they were removed from and re-added to the NIC
(almost) every refresh interval.