Bump default refresh timeout to 15s
On Exoscale, we need to fetch all instance details to determine whether
we need to detach an EIP from another instance in the refresh loop. This
can easily exceed the previous default 10s refresh timeout.

Since there's no clean way to do a custom default timeout just for
Exoscale, we simply bump the default timeout to 15s for all providers.